Update patch set 1
Patch Set 1:

> technically, if it's shipped in -lang, we can drop it from all
 > packages depending on -lang.

Even though strictly speaking it should be ok keeping it just in -lang, I'd prefer to have it on every subpackage.

When searching for the package license on an installed system, I thinkit's a bit weird to have to look for it in the -lang subpackage directory.

Patch-set: 1