Remove suds-jurko from requirements.txt
There are two different forks of suds, community-suds and suds-jurko.
Both provide compatible suds but jurko was intended as temporary [1] and
centos9 is moving python-suds to community-suds [2].

We are managing BR on python3-suds manually so we can remove it from
requirements.txt to avoid the mess of autodetected requirement.

[1] https://pypi.org/project/suds-jurko/#history
[2] https://src.fedoraproject.org/rpms/python-suds/c/fa5c70618cecd26e5e000ffc4c8cca49a81b7cb4?branch=rawhide

Change-Id: Id8adb0a380a13a59e16ac9c296be99f9b69bcfd4